+++ src/main/java/lotto/view/InputGuide.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 public enum InputGuide {
     ACCOUNT("구입금액을 입력해 주세요."),
